Move CellPersistable protocols to separate file
Extracted CellRepresentable, CellInitializable, and CellPersistable typealias from Matrix.swift into a new CellPersistable.swift file for better code organization and separation of concerns.
